Subo cambios en parser y modelo. La velocidad del personaje ya no esta harcodeada, modificar en archivo de configuracion para pruebas. Con cualquier velocidad distinta de 1 se ven los problemas de movimiento.

+++ Model/YAMLParser.h
@@ -27,8 +27,8 @@ struct Screen {
 };
 
 struct Configuration {
-	float main_character_speed;
-	float scroll_margin;
+	int main_character_speed;
+	int scroll_margin;
 };
 
 struct Entities {
